Crate objc2_app_kit

Bindings to the AppKit framework

See Apple’s docs and the general docs on framework crates for more information.

Note that a lot of functionality in AppKit requires that the application has initialized properly, which is only done after the application delegate has received applicationDidFinishLaunching.

You should aspire to do all your UI initialization work in there!

NSWindow

Be careful when creating NSWindow if it’s not inside a window controller; in those cases you’re required to call window.releasedWhenClosed(false) to get correct memory management, which is also why the creation methods for NSWindow are unsafe.

Structs

NSImage

Apple’s documentation

NSImageCacheMode

Apple’s documentation

NSImageLoadStatus

Apple’s documentation

NSImageResizingMode

Apple’s documentation

NSImageSymbolColorRenderingMode

Apple’s documentation

NSImageSymbolConfiguration

Apple’s documentation

NSImageSymbolScale

Apple’s documentation

NSImageSymbolVariableValueMode

Apple’s documentation

NSPasteboard

Apple’s documentation

NSPasteboardAccessBehavior

A value indicating pasteboard access behavior.

NSPasteboardContentsOptions

Apple’s documentation

NSPasteboardItem

Apple’s documentation

NSPasteboardReadingOptions

Apple’s documentation

NSPasteboardWritingOptions

NSPasteboardWriting and NSPasteboardReading Protocols **

Statics

NSColorPboardType⚠Deprecated

Apple’s documentation

NSDragPboard⚠Deprecated

Apple’s documentation

NSFileContentsPboardType⚠

Apple’s documentation

NSFilenamesPboardType⚠Deprecated

Apple’s documentation

NSFilesPromisePboardType⚠Deprecated

Apple’s documentation

NSFindPboard⚠Deprecated

Apple’s documentation

NSFontPboard⚠Deprecated

Apple’s documentation

NSFontPboardType⚠Deprecated

Apple’s documentation

NSGeneralPboard⚠Deprecated

Apple’s documentation

NSHTMLPboardType⚠Deprecated

Apple’s documentation

NSImageNameActionTemplate⚠

Apple’s documentation

NSImageNameAddTemplate⚠

Apple’s documentation

NSImageNameAdvanced⚠

Apple’s documentation

NSImageNameApplicationIcon⚠

Apple’s documentation

NSImageNameBluetoothTemplate⚠

Apple’s documentation

NSImageNameBonjour⚠

Apple’s documentation

NSImageNameBookmarksTemplate⚠

Apple’s documentation

NSImageNameCaution⚠

Apple’s documentation

NSImageNameColorPanel⚠

Apple’s documentation

NSImageNameColumnViewTemplate⚠

Apple’s documentation

NSImageNameComputer⚠

Apple’s documentation

NSImageNameDotMac⚠Deprecated

Apple’s documentation

NSImageNameEnterFullScreenTemplate⚠

Apple’s documentation

NSImageNameEveryone⚠

Apple’s documentation

NSImageNameExitFullScreenTemplate⚠

Apple’s documentation

NSImageNameFlowViewTemplate⚠

Apple’s documentation

NSImageNameFolder⚠

Apple’s documentation

NSImageNameFolderBurnable⚠

Apple’s documentation

NSImageNameFolderSmart⚠

Apple’s documentation

NSImageNameFollowLinkFreestandingTemplate⚠

Apple’s documentation

NSImageNameFontPanel⚠

Apple’s documentation

NSImageNameGoBackTemplate⚠

Apple’s documentation

NSImageNameGoForwardTemplate⚠

Apple’s documentation

NSImageNameGoLeftTemplate⚠

Apple’s documentation

NSImageNameGoRightTemplate⚠

Apple’s documentation

NSImageNameHomeTemplate⚠

Apple’s documentation

NSImageNameIChatTheaterTemplate⚠

Apple’s documentation

NSImageNameIconViewTemplate⚠

Apple’s documentation

NSImageNameInfo⚠

Apple’s documentation

NSImageNameInvalidDataFreestandingTemplate⚠

Apple’s documentation

NSImageNameLeftFacingTriangleTemplate⚠

Apple’s documentation

NSImageNameListViewTemplate⚠

Apple’s documentation

NSImageNameLockLockedTemplate⚠

Apple’s documentation

NSImageNameLockUnlockedTemplate⚠

Apple’s documentation

NSImageNameMenuMixedStateTemplate⚠

Apple’s documentation

NSImageNameMenuOnStateTemplate⚠

Apple’s documentation

NSImageNameMobileMe⚠

Apple’s documentation

NSImageNameMultipleDocuments⚠

Apple’s documentation

NSImageNameNetwork⚠

Apple’s documentation

NSImageNamePathTemplate⚠

Apple’s documentation

NSImageNamePreferencesGeneral⚠

Apple’s documentation

NSImageNameQuickLookTemplate⚠

Apple’s documentation

NSImageNameRefreshFreestandingTemplate⚠

Apple’s documentation

NSImageNameRefreshTemplate⚠

Apple’s documentation

NSImageNameRemoveTemplate⚠

Apple’s documentation

NSImageNameRevealFreestandingTemplate⚠

Apple’s documentation

NSImageNameRightFacingTriangleTemplate⚠

Apple’s documentation

NSImageNameShareTemplate⚠

Apple’s documentation

NSImageNameSlideshowTemplate⚠

Apple’s documentation

NSImageNameSmartBadgeTemplate⚠

Apple’s documentation

NSImageNameStatusAvailable⚠

Apple’s documentation

NSImageNameStatusNone⚠

Apple’s documentation

NSImageNameStatusPartiallyAvailable⚠

Apple’s documentation

NSImageNameStatusUnavailable⚠

Apple’s documentation

NSImageNameStopProgressFreestandingTemplate⚠

Apple’s documentation

NSImageNameStopProgressTemplate⚠

Apple’s documentation

NSImageNameTouchBarAddDetailTemplate⚠

Apple’s documentation

NSImageNameTouchBarAddTemplate⚠

Apple’s documentation

NSImageNameTouchBarAlarmTemplate⚠

Apple’s documentation

NSImageNameTouchBarAudioInputMuteTemplate⚠

Apple’s documentation

NSImageNameTouchBarAudioInputTemplate⚠

Apple’s documentation

NSImageNameTouchBarAudioOutputMuteTemplate⚠

Apple’s documentation

NSImageNameTouchBarAudioOutputVolumeHighTemplate⚠

Apple’s documentation

NSImageNameTouchBarAudioOutputVolumeLowTemplate⚠

Apple’s documentation

NSImageNameTouchBarAudioOutputVolumeMediumTemplate⚠

Apple’s documentation

NSImageNameTouchBarAudioOutputVolumeOffTemplate⚠

Apple’s documentation

NSImageNameTouchBarBookmarksTemplate⚠

Apple’s documentation

NSImageNameTouchBarColorPickerFill⚠

Apple’s documentation

NSImageNameTouchBarColorPickerFont⚠

Apple’s documentation

NSImageNameTouchBarColorPickerStroke⚠

Apple’s documentation

NSImageNameTouchBarCommunicationAudioTemplate⚠

Apple’s documentation

NSImageNameTouchBarCommunicationVideoTemplate⚠

Apple’s documentation

NSImageNameTouchBarComposeTemplate⚠

Apple’s documentation

NSImageNameTouchBarDeleteTemplate⚠

Apple’s documentation

NSImageNameTouchBarDownloadTemplate⚠

Apple’s documentation

NSImageNameTouchBarEnterFullScreenTemplate⚠

Apple’s documentation

NSImageNameTouchBarExitFullScreenTemplate⚠

Apple’s documentation

NSImageNameTouchBarFastForwardTemplate⚠

Apple’s documentation

NSImageNameTouchBarFolderCopyToTemplate⚠

Apple’s documentation

NSImageNameTouchBarFolderMoveToTemplate⚠

Apple’s documentation

NSImageNameTouchBarFolderTemplate⚠

Apple’s documentation

NSImageNameTouchBarGetInfoTemplate⚠

Apple’s documentation

NSImageNameTouchBarGoBackTemplate⚠

Apple’s documentation

NSImageNameTouchBarGoDownTemplate⚠

Apple’s documentation

NSImageNameTouchBarGoForwardTemplate⚠

Apple’s documentation

NSImageNameTouchBarGoUpTemplate⚠

Apple’s documentation

NSImageNameTouchBarHistoryTemplate⚠

Apple’s documentation

NSImageNameTouchBarIconViewTemplate⚠

Apple’s documentation

NSImageNameTouchBarListViewTemplate⚠

Apple’s documentation

NSImageNameTouchBarMailTemplate⚠

Apple’s documentation

NSImageNameTouchBarNewFolderTemplate⚠

Apple’s documentation

NSImageNameTouchBarNewMessageTemplate⚠

Apple’s documentation

NSImageNameTouchBarOpenInBrowserTemplate⚠

Apple’s documentation

NSImageNameTouchBarPauseTemplate⚠

Apple’s documentation

NSImageNameTouchBarPlayPauseTemplate⚠

Apple’s documentation

NSImageNameTouchBarPlayTemplate⚠

Apple’s documentation

NSImageNameTouchBarPlayheadTemplate⚠

Apple’s documentation

NSImageNameTouchBarQuickLookTemplate⚠

Apple’s documentation

NSImageNameTouchBarRecordStartTemplate⚠

Apple’s documentation

NSImageNameTouchBarRecordStopTemplate⚠

Apple’s documentation

NSImageNameTouchBarRefreshTemplate⚠

Apple’s documentation

NSImageNameTouchBarRemoveTemplate⚠

Apple’s documentation

NSImageNameTouchBarRewindTemplate⚠

Apple’s documentation

NSImageNameTouchBarRotateLeftTemplate⚠

Apple’s documentation

NSImageNameTouchBarRotateRightTemplate⚠

Apple’s documentation

NSImageNameTouchBarSearchTemplate⚠

Apple’s documentation

NSImageNameTouchBarShareTemplate⚠

Apple’s documentation

NSImageNameTouchBarSidebarTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipAhead15SecondsTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipAhead30SecondsTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipAheadTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipBack15SecondsTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipBack30SecondsTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipBackTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipToEndTemplate⚠

Apple’s documentation

NSImageNameTouchBarSkipToStartTemplate⚠

Apple’s documentation

NSImageNameTouchBarSlideshowTemplate⚠

Apple’s documentation

NSImageNameTouchBarTagIconTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextBoldTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextBoxTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextCenterAlignTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextItalicTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextJustifiedAlignTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextLeftAlignTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextListTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextRightAlignTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextStrikethroughTemplate⚠

Apple’s documentation

NSImageNameTouchBarTextUnderlineTemplate⚠

Apple’s documentation

NSImageNameTouchBarUserAddTemplate⚠

Apple’s documentation

NSImageNameTouchBarUserGroupTemplate⚠

Apple’s documentation

NSImageNameTouchBarUserTemplate⚠

Apple’s documentation

NSImageNameTouchBarVolumeDownTemplate⚠

Apple’s documentation

NSImageNameTouchBarVolumeUpTemplate⚠

Apple’s documentation

NSImageNameTrashEmpty⚠

Apple’s documentation

NSImageNameTrashFull⚠

Apple’s documentation

NSImageNameUser⚠

Apple’s documentation

NSImageNameUserAccounts⚠

Apple’s documentation

NSImageNameUserGroup⚠

Apple’s documentation

NSImageNameUserGuest⚠

Apple’s documentation

NSInkTextPboardType⚠Deprecated

Apple’s documentation

NSMultipleTextSelectionPboardType⚠Deprecated

Apple’s documentation

NSPDFPboardType⚠Deprecated

Apple’s documentation

NSPICTPboardType⚠Deprecated

Apple’s documentation

NSPasteboardDetectionPatternCalendarEvent⚠

A pattern that indicates the pasteboard detects a string that contains a calendar event.

NSPasteboardDetectionPatternEmailAddress⚠

A pattern that indicates the pasteboard detects a string that contains an email address.

NSPasteboardDetectionPatternFlightNumber⚠

A pattern that indicates the pasteboard detects a string that contains a flight number.

NSPasteboardDetectionPatternLink⚠

A pattern that indicates the pasteboard detects a string that contains a URL.

NSPasteboardDetectionPatternMoneyAmount⚠

A pattern that indicates the pasteboard detects a string that contains an amount of money.

NSPasteboardDetectionPatternNumber⚠

A pattern that indicates the pasteboard detects a string that consists of a numeric value.

NSPasteboardDetectionPatternPhoneNumber⚠

A pattern that indicates the pasteboard detects a string that contains a phone number.

NSPasteboardDetectionPatternPostalAddress⚠

A pattern that indicates the pasteboard detects a string that contains a postal address.

NSPasteboardDetectionPatternProbableWebSearch⚠

A pattern that indicates the pasteboard detects a string suitable for use as a web search term.

NSPasteboardDetectionPatternProbableWebURL⚠

A pattern that indicates the pasteboard detects a string that consists of a web URL.

NSPasteboardDetectionPatternShipmentTrackingNumber⚠

A pattern that indicates the pasteboard detects a string that contains a parcel tracking number and carrier.

NSPasteboardMetadataTypeContentType⚠

A metadata type that returns the content type if the pasteboard detects a reference to a file.

NSPasteboardNameDrag⚠

Apple’s documentation

NSPasteboardNameFind⚠

Apple’s documentation

NSPasteboardNameFont⚠

Apple’s documentation

NSPasteboardNameGeneral⚠

Apple’s documentation

NSPasteboardNameRuler⚠

Apple’s documentation

NSPasteboardTypeColor⚠

Apple’s documentation

NSPasteboardTypeFileURL⚠

Apple’s documentation

NSPasteboardTypeFindPanelSearchOptions⚠Deprecated

Apple’s documentation

NSPasteboardTypeFont⚠

Apple’s documentation

NSPasteboardTypeHTML⚠

Apple’s documentation

NSPasteboardTypeMultipleTextSelection⚠

Apple’s documentation

NSPasteboardTypePDF⚠

Apple’s documentation

NSPasteboardTypePNG⚠

Apple’s documentation

NSPasteboardTypeRTF⚠

Apple’s documentation

NSPasteboardTypeRTFD⚠

Apple’s documentation

NSPasteboardTypeRuler⚠

Apple’s documentation

NSPasteboardTypeSound⚠

Apple’s documentation

NSPasteboardTypeString⚠

Apple’s documentation

NSPasteboardTypeTIFF⚠

Apple’s documentation

NSPasteboardTypeTabularText⚠

Apple’s documentation

NSPasteboardTypeTextFinderOptions⚠

Apple’s documentation

NSPasteboardTypeURL⚠

Apple’s documentation

NSPasteboardURLReadingContentsConformToTypesKey⚠

Apple’s documentation

NSPasteboardURLReadingFileURLsOnlyKey⚠

Apple’s documentation

NSPostScriptPboardType⚠Deprecated

Apple’s documentation

NSRTFDPboardType⚠Deprecated

Apple’s documentation

NSRTFPboardType⚠Deprecated

Apple’s documentation

NSRulerPboard⚠Deprecated

Apple’s documentation

NSRulerPboardType⚠Deprecated

Apple’s documentation

NSStringPboardType⚠Deprecated

Deprecated **

NSTIFFPboardType⚠Deprecated

Apple’s documentation

NSTabularTextPboardType⚠Deprecated

Apple’s documentation

NSURLPboardType⚠Deprecated

Apple’s documentation

NSVCardPboardType⚠Deprecated

Apple’s documentation

Traits

NSBundleImageExtension

Category on NSBundle.

NSImageDelegate

Apple’s documentation

NSPasteboardItemDataProvider

Apple’s documentation

NSPasteboardReading

Apple’s documentation

NSPasteboardTypeOwner

Apple’s documentation

NSPasteboardWriting

Apple’s documentation

NSURLNSPasteboardSupport

Category “NSPasteboardSupport” on NSURL.

Functions

NSCreateFileContentsPboardType

NSCreateFilenamePboardType

NSGetFileType

NSGetFileTypes

Type Aliases

NSImageName

Apple’s documentation

NSPasteboardDetectionPattern

A pattern to detect on the pasteboard, such as a URL, text, or a number.

NSPasteboardMetadataType

A metadata type to detect on the pasteboard.

NSPasteboardName

Apple’s documentation

NSPasteboardReadingOptionKey

Apple’s documentation

NSPasteboardType

Apple’s documentation