Struct objc2_app_kit::NSColor

#[repr(C)]
pub struct NSColor { /* private fields */ }

Implementations

impl NSColor

pub unsafe fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

pub unsafe fn colorWithSRGBRed_green_blue_alpha( red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithGenericGamma22White_alpha( white: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithDisplayP3Red_green_blue_alpha( red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithWhite_alpha( white: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithRed_green_blue_alpha( red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithHue_saturation_brightness_alpha( hue: CGFloat, saturation: CGFloat, brightness: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithDeviceWhite_alpha( white: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithDeviceRed_green_blue_alpha( red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithDeviceHue_saturation_brightness_alpha( hue: CGFloat, saturation: CGFloat, brightness: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithDeviceCyan_magenta_yellow_black_alpha( cyan: CGFloat, magenta: CGFloat, yellow: CGFloat, black: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithCalibratedWhite_alpha( white: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithCalibratedRed_green_blue_alpha( red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithCalibratedHue_saturation_brightness_alpha( hue: CGFloat, saturation: CGFloat, brightness: CGFloat, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn colorWithPatternImage(image: &NSImage) -> Retained<NSColor>

pub unsafe fn type(&self) -> NSColorType

pub unsafe fn colorUsingType( &self, type: NSColorType, ) -> Option<Retained<NSColor>>

pub unsafe fn blackColor() -> Retained<NSColor>

pub unsafe fn darkGrayColor() -> Retained<NSColor>

pub unsafe fn lightGrayColor() -> Retained<NSColor>

pub unsafe fn whiteColor() -> Retained<NSColor>

pub unsafe fn grayColor() -> Retained<NSColor>

pub unsafe fn redColor() -> Retained<NSColor>

pub unsafe fn greenColor() -> Retained<NSColor>

pub unsafe fn blueColor() -> Retained<NSColor>

pub unsafe fn cyanColor() -> Retained<NSColor>

pub unsafe fn yellowColor() -> Retained<NSColor>

pub unsafe fn magentaColor() -> Retained<NSColor>

pub unsafe fn orangeColor() -> Retained<NSColor>

pub unsafe fn purpleColor() -> Retained<NSColor>

pub unsafe fn brownColor() -> Retained<NSColor>

pub unsafe fn clearColor() -> Retained<NSColor>

pub unsafe fn labelColor() -> Retained<NSColor>

pub unsafe fn secondaryLabelColor() -> Retained<NSColor>

pub unsafe fn tertiaryLabelColor() -> Retained<NSColor>

pub unsafe fn quaternaryLabelColor() -> Retained<NSColor>

pub unsafe fn quinaryLabelColor() -> Retained<NSColor>

pub unsafe fn linkColor() -> Retained<NSColor>

pub unsafe fn placeholderTextColor() -> Retained<NSColor>

pub unsafe fn windowFrameTextColor() -> Retained<NSColor>

pub unsafe fn selectedMenuItemTextColor() -> Retained<NSColor>

pub unsafe fn alternateSelectedControlTextColor() -> Retained<NSColor>

pub unsafe fn headerTextColor() -> Retained<NSColor>

pub unsafe fn separatorColor() -> Retained<NSColor>

pub unsafe fn gridColor() -> Retained<NSColor>

pub unsafe fn windowBackgroundColor() -> Retained<NSColor>

pub unsafe fn underPageBackgroundColor() -> Retained<NSColor>

pub unsafe fn controlBackgroundColor() -> Retained<NSColor>

pub unsafe fn selectedContentBackgroundColor() -> Retained<NSColor>

pub unsafe fn unemphasizedSelectedContentBackgroundColor() -> Retained<NSColor>

pub unsafe fn alternatingContentBackgroundColors() -> Retained<NSArray<NSColor>>

pub unsafe fn findHighlightColor() -> Retained<NSColor>

pub unsafe fn textColor() -> Retained<NSColor>

pub unsafe fn textBackgroundColor() -> Retained<NSColor>

pub unsafe fn textInsertionPointColor() -> Retained<NSColor>

pub unsafe fn selectedTextColor() -> Retained<NSColor>

pub unsafe fn selectedTextBackgroundColor() -> Retained<NSColor>

pub unsafe fn unemphasizedSelectedTextBackgroundColor() -> Retained<NSColor>

pub unsafe fn unemphasizedSelectedTextColor() -> Retained<NSColor>

pub unsafe fn controlColor() -> Retained<NSColor>

pub unsafe fn controlTextColor() -> Retained<NSColor>

pub unsafe fn selectedControlColor() -> Retained<NSColor>

pub unsafe fn selectedControlTextColor() -> Retained<NSColor>

pub unsafe fn disabledControlTextColor() -> Retained<NSColor>

pub unsafe fn keyboardFocusIndicatorColor() -> Retained<NSColor>

pub unsafe fn scrubberTexturedBackgroundColor() -> Retained<NSColor>

pub unsafe fn systemRedColor() -> Retained<NSColor>

pub unsafe fn systemGreenColor() -> Retained<NSColor>

pub unsafe fn systemBlueColor() -> Retained<NSColor>

pub unsafe fn systemOrangeColor() -> Retained<NSColor>

pub unsafe fn systemYellowColor() -> Retained<NSColor>

pub unsafe fn systemBrownColor() -> Retained<NSColor>

pub unsafe fn systemPinkColor() -> Retained<NSColor>

pub unsafe fn systemPurpleColor() -> Retained<NSColor>

pub unsafe fn systemGrayColor() -> Retained<NSColor>

pub unsafe fn systemTealColor() -> Retained<NSColor>

pub unsafe fn systemIndigoColor() -> Retained<NSColor>

pub unsafe fn systemMintColor() -> Retained<NSColor>

pub unsafe fn systemCyanColor() -> Retained<NSColor>

pub unsafe fn systemFillColor() -> Retained<NSColor>

pub unsafe fn secondarySystemFillColor() -> Retained<NSColor>

pub unsafe fn tertiarySystemFillColor() -> Retained<NSColor>

pub unsafe fn quaternarySystemFillColor() -> Retained<NSColor>

pub unsafe fn quinarySystemFillColor() -> Retained<NSColor>

pub unsafe fn controlAccentColor() -> Retained<NSColor>

pub unsafe fn currentControlTint() -> NSControlTint

pub unsafe fn colorForControlTint( control_tint: NSControlTint, ) -> Retained<NSColor>

👎Deprecated: NSControlTint does not describe the full range of available control accent colors. Use +[NSColor controlAccentColor] instead.

pub unsafe fn highlightColor() -> Retained<NSColor>

pub unsafe fn shadowColor() -> Retained<NSColor>

pub unsafe fn highlightWithLevel( &self, val: CGFloat, ) -> Option<Retained<NSColor>>

pub unsafe fn shadowWithLevel(&self, val: CGFloat) -> Option<Retained<NSColor>>

pub unsafe fn colorWithSystemEffect( &self, system_effect: NSColorSystemEffect, ) -> Retained<NSColor>

pub unsafe fn set(&self)

pub unsafe fn setFill(&self)

pub unsafe fn setStroke(&self)

pub unsafe fn blendedColorWithFraction_ofColor( &self, fraction: CGFloat, color: &NSColor, ) -> Option<Retained<NSColor>>

pub unsafe fn colorWithAlphaComponent( &self, alpha: CGFloat, ) -> Retained<NSColor>

pub unsafe fn localizedCatalogNameComponent(&self) -> Retained<NSString>

pub unsafe fn localizedColorNameComponent(&self) -> Retained<NSString>

pub unsafe fn redComponent(&self) -> CGFloat

pub unsafe fn greenComponent(&self) -> CGFloat

pub unsafe fn blueComponent(&self) -> CGFloat

pub unsafe fn getRed_green_blue_alpha( &self, red: *mut CGFloat, green: *mut CGFloat, blue: *mut CGFloat, alpha: *mut CGFloat, )

pub unsafe fn hueComponent(&self) -> CGFloat

pub unsafe fn saturationComponent(&self) -> CGFloat

pub unsafe fn brightnessComponent(&self) -> CGFloat

pub unsafe fn getHue_saturation_brightness_alpha( &self, hue: *mut CGFloat, saturation: *mut CGFloat, brightness: *mut CGFloat, alpha: *mut CGFloat, )

pub unsafe fn whiteComponent(&self) -> CGFloat

pub unsafe fn getWhite_alpha(&self, white: *mut CGFloat, alpha: *mut CGFloat)

pub unsafe fn cyanComponent(&self) -> CGFloat

pub unsafe fn magentaComponent(&self) -> CGFloat

pub unsafe fn yellowComponent(&self) -> CGFloat

pub unsafe fn blackComponent(&self) -> CGFloat

pub unsafe fn getCyan_magenta_yellow_black_alpha( &self, cyan: *mut CGFloat, magenta: *mut CGFloat, yellow: *mut CGFloat, black: *mut CGFloat, alpha: *mut CGFloat, )

pub unsafe fn numberOfComponents(&self) -> NSInteger

pub unsafe fn getComponents(&self, components: NonNull<CGFloat>)

pub unsafe fn patternImage(&self) -> Retained<NSImage>

pub unsafe fn alphaComponent(&self) -> CGFloat

pub unsafe fn colorFromPasteboard( paste_board: &NSPasteboard, ) -> Option<Retained<NSColor>>

pub unsafe fn writeToPasteboard(&self, paste_board: &NSPasteboard)

pub unsafe fn drawSwatchInRect(&self, rect: NSRect)

pub unsafe fn ignoresAlpha(mtm: MainThreadMarker) -> bool

👎Deprecated: Use showsAlpha in NSColorPanel and supportsAlpha in NSColorWell to control alpha behavior for individual controls.

pub unsafe fn setIgnoresAlpha(ignores_alpha: bool, mtm: MainThreadMarker)

👎Deprecated: Use showsAlpha in NSColorPanel and supportsAlpha in NSColorWell to control alpha behavior for individual controls.

impl NSColor

Methods declared on superclass NSObject

pub unsafe fn new() -> Retained<Self>

impl NSColor

NSDeprecated

pub unsafe fn controlHighlightColor() -> Retained<NSColor>

👎Deprecated: Use a color that matches the semantics being used, such as separatorColor

pub unsafe fn controlLightHighlightColor() -> Retained<NSColor>

👎Deprecated: Use a color that matches the semantics being used, such as separatorColor

pub unsafe fn controlShadowColor() -> Retained<NSColor>

👎Deprecated: Use a color that matches the semantics being used, such as separatorColor

pub unsafe fn controlDarkShadowColor() -> Retained<NSColor>

👎Deprecated: Use a color that matches the semantics being used, such as separatorColor

pub unsafe fn scrollBarColor() -> Retained<NSColor>

👎Deprecated: Use NSScroller instead

pub unsafe fn knobColor() -> Retained<NSColor>

👎Deprecated: Use NSScroller instead

pub unsafe fn selectedKnobColor() -> Retained<NSColor>

👎Deprecated: Use NSScroller instead

pub unsafe fn windowFrameColor() -> Retained<NSColor>

👎Deprecated: Use NSVisualEffectMaterialTitlebar

pub unsafe fn selectedMenuItemColor() -> Retained<NSColor>

👎Deprecated: Use NSVisualEffectMaterialSelection

pub unsafe fn headerColor() -> Retained<NSColor>

👎Deprecated: Use NSVisualEffectMaterialHeaderView

pub unsafe fn secondarySelectedControlColor() -> Retained<NSColor>

👎Deprecated

pub unsafe fn alternateSelectedControlColor() -> Retained<NSColor>

👎Deprecated

pub unsafe fn controlAlternatingRowBackgroundColors() -> Retained<NSArray<NSColor>>

👎Deprecated

pub unsafe fn colorSpaceName(&self) -> Retained<NSColorSpaceName>

👎Deprecated: Use -type and NSColorType instead

pub unsafe fn colorUsingColorSpaceName_device( &self, name: Option<&NSColorSpaceName>, device_description: Option<&NSDictionary<NSDeviceDescriptionKey, AnyObject>>, ) -> Option<Retained<NSColor>>

👎Deprecated: Use -colorUsingType: or -colorUsingColorSpace: instead

pub unsafe fn colorUsingColorSpaceName( &self, name: &NSColorSpaceName, ) -> Option<Retained<NSColor>>

👎Deprecated: Use -colorUsingType: or -colorUsingColorSpace: instead

impl NSColor

This impl block contains no items.

NSQuartzCoreAdditions