Struct objc2_app_kit::NSImage

#[repr(C)]
pub struct NSImage { /* private fields */ }

Implementations

impl NSImage

pub unsafe fn imageNamed(name: &NSImageName) -> Option<Retained<NSImage>>

pub unsafe fn imageWithSystemSymbolName_accessibilityDescription( name: &NSString, description: Option<&NSString>, ) -> Option<Retained<Self>>

pub unsafe fn imageWithSystemSymbolName_variableValue_accessibilityDescription( name: &NSString, value: c_double, description: Option<&NSString>, ) -> Option<Retained<Self>>

pub unsafe fn imageWithSymbolName_variableValue( name: &NSString, value: c_double, ) -> Option<Retained<Self>>

pub unsafe fn imageWithSymbolName_bundle_variableValue( name: &NSString, bundle: Option<&NSBundle>, value: c_double, ) -> Option<Retained<Self>>

pub unsafe fn initWithSize( this: Allocated<Self>, size: NSSize, ) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Retained<Self>

pub fn initWithData( this: Allocated<Self>, data: &NSData, ) -> Option<Retained<Self>>

pub unsafe fn initWithContentsOfFile( this: Allocated<Self>, file_name: &NSString, ) -> Option<Retained<Self>>

pub unsafe fn initWithContentsOfURL( this: Allocated<Self>, url: &NSURL, ) -> Option<Retained<Self>>

pub fn initByReferencingFile( this: Allocated<Self>, file_name: &NSString, ) -> Option<Retained<Self>>

pub unsafe fn initByReferencingURL( this: Allocated<Self>, url: &NSURL, ) -> Retained<Self>

pub unsafe fn initWithPasteboard( this: Allocated<Self>, pasteboard: &NSPasteboard, ) -> Option<Retained<Self>>

pub unsafe fn initWithDataIgnoringOrientation( this: Allocated<Self>, data: &NSData, ) -> Option<Retained<Self>>

pub unsafe fn imageWithSize_flipped_drawingHandler( size: NSSize, drawing_handler_should_be_called_with_flipped_context: bool, drawing_handler: &Block<dyn Fn(NSRect) -> Bool>, ) -> Retained<Self>

pub unsafe fn size(&self) -> NSSize

pub unsafe fn setSize(&self, size: NSSize)

pub unsafe fn setName(&self, string: Option<&NSImageName>) -> bool

pub unsafe fn name(&self) -> Option<Retained<NSImageName>>

pub unsafe fn backgroundColor(&self) -> Retained<NSColor>

pub unsafe fn setBackgroundColor(&self, background_color: &NSColor)

pub unsafe fn usesEPSOnResolutionMismatch(&self) -> bool

pub unsafe fn setUsesEPSOnResolutionMismatch( &self, uses_eps_on_resolution_mismatch: bool, )

pub unsafe fn prefersColorMatch(&self) -> bool

pub unsafe fn setPrefersColorMatch(&self, prefers_color_match: bool)

pub unsafe fn matchesOnMultipleResolution(&self) -> bool

pub unsafe fn setMatchesOnMultipleResolution( &self, matches_on_multiple_resolution: bool, )

pub unsafe fn matchesOnlyOnBestFittingAxis(&self) -> bool

pub unsafe fn setMatchesOnlyOnBestFittingAxis( &self, matches_only_on_best_fitting_axis: bool, )

pub unsafe fn drawAtPoint_fromRect_operation_fraction( &self, point: NSPoint, from_rect: NSRect, op: NSCompositingOperation, delta: CGFloat, )

pub unsafe fn drawInRect_fromRect_operation_fraction( &self, rect: NSRect, from_rect: NSRect, op: NSCompositingOperation, delta: CGFloat, )

pub unsafe fn drawInRect(&self, rect: NSRect)

pub unsafe fn recache(&self)

pub unsafe fn TIFFRepresentation(&self) -> Option<Retained<NSData>>

pub unsafe fn isValid(&self) -> bool

pub unsafe fn delegate( &self, ) -> Option<Retained<ProtocolObject<dyn NSImageDelegate>>>

pub unsafe fn setDelegate( &self, delegate: Option<&ProtocolObject<dyn NSImageDelegate>>, )

pub unsafe fn imageTypes() -> Retained<NSArray<NSString>>

pub unsafe fn imageUnfilteredTypes() -> Retained<NSArray<NSString>>

pub unsafe fn canInitWithPasteboard(pasteboard: &NSPasteboard) -> bool

pub unsafe fn cacheMode(&self) -> NSImageCacheMode

pub unsafe fn setCacheMode(&self, cache_mode: NSImageCacheMode)

pub unsafe fn alignmentRect(&self) -> NSRect

pub unsafe fn setAlignmentRect(&self, alignment_rect: NSRect)

pub unsafe fn isTemplate(&self) -> bool

pub unsafe fn setTemplate(&self, template: bool)

pub unsafe fn accessibilityDescription(&self) -> Option<Retained<NSString>>

pub unsafe fn setAccessibilityDescription( &self, accessibility_description: Option<&NSString>, )

pub unsafe fn recommendedLayerContentsScale( &self, preferred_contents_scale: CGFloat, ) -> CGFloat

pub unsafe fn layerContentsForContentsScale( &self, layer_contents_scale: CGFloat, ) -> Retained<AnyObject>

pub unsafe fn capInsets(&self) -> NSEdgeInsets

pub unsafe fn setCapInsets(&self, cap_insets: NSEdgeInsets)

pub unsafe fn resizingMode(&self) -> NSImageResizingMode

pub unsafe fn setResizingMode(&self, resizing_mode: NSImageResizingMode)

pub unsafe fn imageWithSymbolConfiguration( &self, configuration: &NSImageSymbolConfiguration, ) -> Option<Retained<NSImage>>

pub unsafe fn symbolConfiguration(&self) -> Retained<NSImageSymbolConfiguration>

pub unsafe fn imageWithLocale( &self, locale: Option<&NSLocale>, ) -> Retained<NSImage>

pub unsafe fn locale(&self) -> Option<Retained<NSLocale>>

impl NSImage

Methods declared on superclass NSObject

pub unsafe fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn new() -> Retained<Self>

impl NSImage

Deprecated

pub unsafe fn imageUnfilteredFileTypes() -> Retained<NSArray<NSString>>

👎Deprecated: Use +imageUnfilteredTypes instead

pub unsafe fn imageUnfilteredPasteboardTypes() -> Retained<NSArray<NSPasteboardType>>

👎Deprecated: Use +imageUnfilteredTypes instead

pub unsafe fn imageFileTypes() -> Retained<NSArray<NSString>>

👎Deprecated: Use +imageTypes instead

pub unsafe fn imagePasteboardTypes() -> Retained<NSArray<NSPasteboardType>>

👎Deprecated: Use +imageTypes instead

pub unsafe fn lockFocus(&self)

👎Deprecated: This method is incompatible with resolution-independent drawing and should not be used. Instead, use +[NSImage imageWithSize:flipped:drawingHandler:] to create a block-based image describing the desired image drawing, or use +[NSGraphicsContext graphicsContextWithBitmapImageRep:] to manipulate specific bitmap image representations.

pub unsafe fn lockFocusFlipped(&self, flipped: bool)

👎Deprecated: This method is incompatible with resolution-independent drawing and should not be used. Instead, use +[NSImage imageWithSize:flipped:drawingHandler:] to create a block-based image describing the desired image drawing, or use +[NSGraphicsContext graphicsContextWithBitmapImageRep:] to manipulate specific bitmap image representations.

pub unsafe fn unlockFocus(&self)

👎Deprecated: This method is incompatible with resolution-independent drawing and should not be used. Instead, use +[NSImage imageWithSize:flipped:drawingHandler:] to create a block-based image describing the desired image drawing, or use +[NSGraphicsContext graphicsContextWithBitmapImageRep:] to manipulate specific bitmap image representations.

pub unsafe fn setFlipped(&self, flag: bool)

👎Deprecated: The concept of flippedness for NSImage is deprecated. Please see the AppKit 10.6 release notes for a discussion of why and for how to replace existing usage.

pub unsafe fn isFlipped(&self) -> bool

👎Deprecated: The concept of flippedness for NSImage is deprecated. Please see the AppKit 10.6 release notes for a discussion of why and for how to replace existing usage.

pub unsafe fn setScalesWhenResized(&self, flag: bool)

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn scalesWhenResized(&self) -> bool

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn setDataRetained(&self, flag: bool)

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn isDataRetained(&self) -> bool

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn setCachedSeparately(&self, flag: bool)

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn isCachedSeparately(&self) -> bool

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn setCacheDepthMatchesImageDepth(&self, flag: bool)

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn cacheDepthMatchesImageDepth(&self) -> bool

👎Deprecated: You should be able to remove use of this method without any replacement. See 10.6 AppKit release notes for details.

pub unsafe fn dissolveToPoint_fraction(&self, point: NSPoint, fraction: CGFloat)

👎Deprecated: Use -drawAtPoint:… or -drawInRect:… methods instead

pub unsafe fn dissolveToPoint_fromRect_fraction( &self, point: NSPoint, rect: NSRect, fraction: CGFloat, )

👎Deprecated: Use -drawAtPoint:… or -drawInRect:… methods instead

pub unsafe fn compositeToPoint_operation( &self, point: NSPoint, operation: NSCompositingOperation, )

👎Deprecated: Use -drawAtPoint:… or -drawInRect:… methods instead

pub unsafe fn compositeToPoint_fromRect_operation( &self, point: NSPoint, rect: NSRect, operation: NSCompositingOperation, )

👎Deprecated: Use -drawAtPoint:… or -drawInRect:… methods instead

pub unsafe fn compositeToPoint_operation_fraction( &self, point: NSPoint, operation: NSCompositingOperation, fraction: CGFloat, )

👎Deprecated: Use -drawAtPoint:… or -drawInRect:… methods instead

pub unsafe fn compositeToPoint_fromRect_operation_fraction( &self, point: NSPoint, rect: NSRect, operation: NSCompositingOperation, fraction: CGFloat, )

👎Deprecated: Use -drawAtPoint:… or -drawInRect:… methods instead

pub unsafe fn cancelIncrementalLoad(&self)

👎Deprecated: This method does not perform any operation on 10.4 or later.