Crate objc2_app_kit

Bindings to the AppKit framework

See Apple’s docs and the general docs on framework crates for more information.

Note that a lot of functionality in AppKit requires that the application has initialized properly, which is only done after the application delegate has received applicationDidFinishLaunching.

You should aspire to do all your UI initialization work in there!

NSWindow

Be careful when creating NSWindow if it’s not inside a window controller; in those cases you’re required to call window.releasedWhenClosed(false) to get correct memory management, which is also why the creation methods for NSWindow are unsafe.

Examples

Implementing NSApplicationDelegate for a custom class.

#![deny(unsafe_op_in_unsafe_fn)]
use objc2::rc::Retained;
use objc2::runtime::ProtocolObject;
use objc2::{declare_class, msg_send_id, mutability, ClassType, DeclaredClass};
use objc2_app_kit::{NSApplication, NSApplicationActivationPolicy, NSApplicationDelegate};
use objc2_foundation::{
    ns_string, MainThreadMarker, NSCopying, NSNotification, NSObject, NSObjectProtocol, NSString,
};

#[derive(Debug)]
#[allow(unused)]
struct Ivars {
    ivar: u8,
    another_ivar: bool,
    box_ivar: Box<i32>,
    maybe_box_ivar: Option<Box<i32>>,
    id_ivar: Retained<NSString>,
    maybe_id_ivar: Option<Retained<NSString>>,
}

declare_class!(
    struct AppDelegate;

    // SAFETY:
    // - The superclass NSObject does not have any subclassing requirements.
    // - Main thread only mutability is correct, since this is an application delegate.
    // - `AppDelegate` does not implement `Drop`.
    unsafe impl ClassType for AppDelegate {
        type Super = NSObject;
        type Mutability = mutability::MainThreadOnly;
        const NAME: &'static str = "MyAppDelegate";
    }

    impl DeclaredClass for AppDelegate {
        type Ivars = Ivars;
    }

    unsafe impl NSObjectProtocol for AppDelegate {}

    unsafe impl NSApplicationDelegate for AppDelegate {
        #[method(applicationDidFinishLaunching:)]
        fn did_finish_launching(&self, notification: &NSNotification) {
            println!("Did finish launching!");
            // Do something with the notification
            dbg!(notification);
        }

        #[method(applicationWillTerminate:)]
        fn will_terminate(&self, _notification: &NSNotification) {
            println!("Will terminate!");
        }
    }
);

impl AppDelegate {
    fn new(ivar: u8, another_ivar: bool, mtm: MainThreadMarker) -> Retained<Self> {
        let this = mtm.alloc();
        let this = this.set_ivars(Ivars {
            ivar,
            another_ivar,
            box_ivar: Box::new(2),
            maybe_box_ivar: None,
            id_ivar: NSString::from_str("abc"),
            maybe_id_ivar: Some(ns_string!("def").copy()),
        });
        unsafe { msg_send_id![super(this), init] }
    }
}

fn main() {
    let mtm: MainThreadMarker = MainThreadMarker::new().unwrap();

    let app = NSApplication::sharedApplication(mtm);
    app.setActivationPolicy(NSApplicationActivationPolicy::Regular);

    // configure the application delegate
    let delegate = AppDelegate::new(42, true, mtm);
    let object = ProtocolObject::from_ref(&*delegate);
    app.setDelegate(Some(object));

    // run the app
    unsafe { app.run() };
}

An example showing basic and a bit more advanced usage of NSPasteboard.

//! Read from the global pasteboard, and write a new string into it.
//!
//! Works on macOS 10.7+
#![deny(unsafe_op_in_unsafe_fn)]

use objc2::rc::Retained;
use objc2::runtime::{AnyClass, AnyObject, ProtocolObject};
use objc2::ClassType;
use objc2_app_kit::{NSPasteboard, NSPasteboardTypeString};
use objc2_foundation::{NSArray, NSCopying, NSString};

/// Simplest implementation
pub fn get_text_1(pasteboard: &NSPasteboard) -> Option<Retained<NSString>> {
    unsafe { pasteboard.stringForType(NSPasteboardTypeString) }
}

/// More complex implementation using `readObjectsForClasses:options:`,
/// intended to show how some patterns might require more knowledge of
/// nitty-gritty details.
pub fn get_text_2(pasteboard: &NSPasteboard) -> Option<Retained<NSString>> {
    // The NSPasteboard API is a bit weird, it requires you to pass classes as
    // objects, which `objc2_foundation::NSArray` was not really made for - so
    // we convert the class to an `AnyObject` type instead.
    //
    // TODO: Investigate and find a better way to express this in `objc2`.
    let string_class = {
        let cls: *const AnyClass = NSString::class();
        let cls = cls as *mut AnyObject;
        unsafe { Retained::retain(cls).unwrap() }
    };
    let class_array = NSArray::from_vec(vec![string_class]);
    let objects = unsafe { pasteboard.readObjectsForClasses_options(&class_array, None) };

    let obj: *const AnyObject = objects?.first()?;
    // And this part is weird as well, since we now have to convert the object
    // into an NSString, which we know it to be since that's what we told
    // `readObjectsForClasses:options:`.
    let obj = obj as *mut NSString;
    Some(unsafe { Retained::retain(obj) }.unwrap())
}

pub fn set_text(pasteboard: &NSPasteboard, text: &NSString) {
    let _ = unsafe { pasteboard.clearContents() };
    let obj = ProtocolObject::from_retained(text.copy());
    let objects = NSArray::from_vec(vec![obj]);
    let res = unsafe { pasteboard.writeObjects(&objects) };
    if !res {
        panic!("Failed writing to pasteboard");
    }
}

fn main() {
    let pasteboard = unsafe { NSPasteboard::generalPasteboard() };
    let impl_1 = get_text_1(&pasteboard);
    let impl_2 = get_text_2(&pasteboard);
    println!("Pasteboard text from implementation 1 was: {impl_1:?}");
    println!("Pasteboard text from implementation 2 was: {impl_2:?}");
    assert_eq!(impl_1, impl_2);

    let s = NSString::from_str("Hello, world!");
    set_text(&pasteboard, &s);
    println!("Now the pasteboard text should be: {s:?}");
    assert_eq!(Some(s), get_text_1(&pasteboard));
}

Structs

NSAnimation

NSAnimationBlockingMode

NSAnimationCurve

NSAnimationEffectDeprecated

NSAppearance

NSApplication

NSApplicationActivationOptions

NSApplicationActivationPolicy

NSApplicationDelegateReply

NSApplicationOcclusionState

NSApplicationPresentationOptions

NSApplicationPrintReply

NSApplicationTerminateReply

NSAutoresizingMaskOptions

NSBackgroundStyle

NSBackingStoreType

NSBorderType

NSCell

NSCellAttribute

NSCellHitResult

NSCellImagePosition

NSCellStyleMask

NSCellType

NSColor

NSColorRenderingIntent

NSColorSystemEffect

NSColorType

NSCompositingOperation

NSControlSize

NSControlTint

NSDisplayGamut

NSDragOperation

NSDraggingContext

NSDraggingFormation

NSDraggingItemEnumerationOptions

NSEvent

NSEventButtonMask

NSEventGestureAxis

NSEventMask

NSEventModifierFlags

NSEventPhase

NSEventSubtype

NSEventSwipeTrackingOptions

NSEventType

NSFocusRingPlacement

NSFocusRingType

NSImage

NSImageCacheMode

NSImageLoadStatus

NSImageResizingMode

NSImageScaling

NSImageSymbolConfiguration

NSImageSymbolScale

NSMenu

NSMenuItem

NSMenuPresentationStyle

NSMenuProperties

NSMenuSelectionMode

NSOpenPanel

NSPanel

NSPaperOrientation

NSPasteboard

NSPasteboardContentsOptions

NSPasteboardReadingOptions

NSPasteboardWritingOptions

NSPointingDeviceType

NSPressureBehavior

NSPrintInfo

NSPrintOperation

NSPrintRenderingQuality

NSPrintingOrientationDeprecated

NSPrintingPageOrder

NSPrintingPaginationMode

NSRemoteNotificationType

NSRequestUserAttentionType

NSResponder

NSRunningApplication

NSSavePanel

NSSelectionDirection

NSSpringLoadingHighlight

NSSpringLoadingOptions

NSTextFinder

NSTextFinderAction

NSTextFinderMatchingType

NSTitlebarSeparatorStyle

NSView

NSViewAnimation

NSViewLayerContentsPlacement

NSViewLayerContentsRedrawPolicy

NSVisualEffectBlendingMode

NSVisualEffectMaterial

NSVisualEffectState

NSVisualEffectView

NSWindow

NSWindowAnimationBehavior

NSWindowBackingLocationDeprecated

NSWindowButton

NSWindowCollectionBehavior

NSWindowDepth

NSWindowListOptions

NSWindowNumberListOptions

NSWindowOcclusionState

NSWindowOrderingMode

NSWindowSharingType

NSWindowStyleMask

NSWindowTabbingMode

NSWindowTitleVisibility

NSWindowToolbarStyle

NSWindowUserTabbingPreference

Constants

NSAlertAlternateReturnDeprecated

NSAlertDefaultReturnDeprecated

NSAlertErrorReturnDeprecated

NSAlertOtherReturnDeprecated

NSAnyTypeDeprecated

NSBeginFunctionKey

NSBreakFunctionKey

NSClearDisplayFunctionKey

NSClearLineFunctionKey

NSDeleteCharFunctionKey

NSDeleteFunctionKey

NSDeleteLineFunctionKey

NSDisplayWindowRunLoopOrdering

NSDoubleTypeDeprecated

NSDownArrowFunctionKey

NSEndFunctionKey

NSExecuteFunctionKey

NSF1FunctionKey

NSF2FunctionKey

NSF3FunctionKey

NSF4FunctionKey

NSF5FunctionKey

NSF6FunctionKey

NSF7FunctionKey

NSF8FunctionKey

NSF9FunctionKey

NSF10FunctionKey

NSF11FunctionKey

NSF12FunctionKey

NSF13FunctionKey

NSF14FunctionKey

NSF15FunctionKey

NSF16FunctionKey

NSF17FunctionKey

NSF18FunctionKey

NSF19FunctionKey

NSF20FunctionKey

NSF21FunctionKey

NSF22FunctionKey

NSF23FunctionKey

NSF24FunctionKey

NSF25FunctionKey

NSF26FunctionKey

NSF27FunctionKey

NSF28FunctionKey

NSF29FunctionKey

NSF30FunctionKey

NSF31FunctionKey

NSF32FunctionKey

NSF33FunctionKey

NSF34FunctionKey

NSF35FunctionKey

NSFindFunctionKey

NSFloatTypeDeprecated

NSHelpFunctionKey

NSHomeFunctionKey

NSInsertCharFunctionKey

NSInsertFunctionKey

NSInsertLineFunctionKey

NSIntTypeDeprecated

NSLeftArrowFunctionKey

NSMenuFunctionKey

NSModeSwitchFunctionKey

NSNextFunctionKey

NSPageDownFunctionKey

NSPageUpFunctionKey

NSPauseFunctionKey

NSPositiveDoubleTypeDeprecated

NSPositiveFloatTypeDeprecated

NSPositiveIntTypeDeprecated

NSPrevFunctionKey

NSPrintFunctionKey

NSPrintScreenFunctionKey

NSRedoFunctionKey

NSResetCursorRectsRunLoopOrdering

NSResetFunctionKey

NSRightArrowFunctionKey

NSRunAbortedResponseDeprecated

NSRunContinuesResponseDeprecated

NSRunStoppedResponseDeprecated

NSScrollLockFunctionKey

NSSelectFunctionKey

NSStopFunctionKey

NSSysReqFunctionKey

NSSystemFunctionKey

NSUndoFunctionKey

NSUpArrowFunctionKey

NSUpdateWindowsRunLoopOrdering

NSUserFunctionKey

Statics

NSAWTEventType

NSAboutPanelOptionApplicationIcon⚠

NSAboutPanelOptionApplicationName⚠

NSAboutPanelOptionApplicationVersion⚠

NSAboutPanelOptionCredits⚠

NSAboutPanelOptionVersion⚠

NSAlphaShiftKeyMask

NSAlternateKeyMask

NSAnimationProgressMark⚠

NSAnimationProgressMarkNotification⚠

NSAnimationTriggerOrderIn⚠

NSAnimationTriggerOrderOut⚠

NSAnyEventMask

NSAppKitDefined

NSAppKitDefinedMask

NSAppKitVersionNumber⚠

NSAppKitVersionNumber10_0

NSAppKitVersionNumber10_1

NSAppKitVersionNumber10_2

NSAppKitVersionNumber10_3

NSAppKitVersionNumber10_4

NSAppKitVersionNumber10_5

NSAppKitVersionNumber10_6

NSAppKitVersionNumber10_7

NSAppKitVersionNumber10_8

NSAppKitVersionNumber10_9

NSAppKitVersionNumber10_2_3

NSAppKitVersionNumber10_3_2

NSAppKitVersionNumber10_3_3

NSAppKitVersionNumber10_3_5

NSAppKitVersionNumber10_3_7

NSAppKitVersionNumber10_3_9

NSAppKitVersionNumber10_4_1

NSAppKitVersionNumber10_4_3

NSAppKitVersionNumber10_4_4

NSAppKitVersionNumber10_4_7

NSAppKitVersionNumber10_5_2

NSAppKitVersionNumber10_5_3

NSAppKitVersionNumber10_7_2

NSAppKitVersionNumber10_7_3

NSAppKitVersionNumber10_7_4

NSAppKitVersionNumber10_10

NSAppKitVersionNumber10_11

NSAppKitVersionNumber10_12

NSAppKitVersionNumber10_13

NSAppKitVersionNumber10_14

NSAppKitVersionNumber10_15

NSAppKitVersionNumber10_10_2

NSAppKitVersionNumber10_10_3

NSAppKitVersionNumber10_10_4

NSAppKitVersionNumber10_10_5

NSAppKitVersionNumber10_10_Max

NSAppKitVersionNumber10_11_1

NSAppKitVersionNumber10_11_2

NSAppKitVersionNumber10_11_3

NSAppKitVersionNumber10_12_1

NSAppKitVersionNumber10_12_2

NSAppKitVersionNumber10_13_1

NSAppKitVersionNumber10_13_2

NSAppKitVersionNumber10_13_4

NSAppKitVersionNumber10_14_1

NSAppKitVersionNumber10_14_2

NSAppKitVersionNumber10_14_3

NSAppKitVersionNumber10_14_4

NSAppKitVersionNumber10_14_5

NSAppKitVersionNumber10_15_1

NSAppKitVersionNumber10_15_2

NSAppKitVersionNumber10_15_3

NSAppKitVersionNumber10_15_4

NSAppKitVersionNumber10_15_5

NSAppKitVersionNumber10_15_6

NSAppKitVersionNumber11_0

NSAppKitVersionNumber11_1

NSAppKitVersionNumber11_2

NSAppKitVersionNumber11_3

NSAppKitVersionNumber11_4

NSAppKitVersionNumber11_5

NSAppKitVersionNumber12_0

NSAppKitVersionNumber12_1

NSAppKitVersionNumber12_2

NSAppKitVersionNumber12_3

NSAppKitVersionNumber12_4

NSAppKitVersionNumber12_5

NSAppKitVersionNumber13_0

NSAppKitVersionNumber13_1

NSAppKitVersionNumber13_2

NSAppKitVersionNumber13_3

NSAppKitVersionNumber13_4

NSAppKitVersionNumber13_5

NSAppKitVersionNumber13_6

NSAppKitVersionNumber14_0

NSAppKitVersionNumber14_1

NSAppKitVersionNumberWithCustomSheetPosition

NSAppKitVersionNumberWithDeferredWindowDisplaySupport

NSAppKitVersionNumberWithPatternColorLeakFix

NSAppearanceNameAccessibilityHighContrastAqua⚠

NSAppearanceNameAccessibilityHighContrastDarkAqua⚠

NSAppearanceNameAccessibilityHighContrastVibrantDark⚠

NSAppearanceNameAccessibilityHighContrastVibrantLight⚠

NSAppearanceNameAqua⚠

NSAppearanceNameDarkAqua⚠

NSAppearanceNameLightContent⚠

NSAppearanceNameVibrantDark⚠

NSAppearanceNameVibrantLight⚠

NSApplicationActivatedEventType

NSApplicationDeactivatedEventType

NSApplicationDefined

NSApplicationDefinedMask

NSApplicationDidBecomeActiveNotification⚠

NSApplicationDidChangeOcclusionStateNotification⚠

NSApplicationDidChangeScreenParametersNotification⚠

NSApplicationDidFinishLaunchingNotification⚠

NSApplicationDidHideNotification⚠

NSApplicationDidResignActiveNotification⚠

NSApplicationDidUnhideNotification⚠

NSApplicationDidUpdateNotification⚠

NSApplicationLaunchIsDefaultLaunchKey⚠

NSApplicationLaunchRemoteNotificationKey⚠

NSApplicationLaunchUserNotificationKey⚠

NSApplicationProtectedDataDidBecomeAvailableNotification⚠

NSApplicationProtectedDataWillBecomeUnavailableNotification⚠

NSApplicationWillBecomeActiveNotification⚠

NSApplicationWillFinishLaunchingNotification⚠

NSApplicationWillHideNotification⚠

NSApplicationWillResignActiveNotification⚠

NSApplicationWillTerminateNotification⚠

NSApplicationWillUnhideNotification⚠

NSApplicationWillUpdateNotification⚠

NSAutoPagination

NSBackgroundStyleDark

NSBackgroundStyleLight

NSBackingPropertyOldColorSpaceKey⚠

NSBackingPropertyOldScaleFactorKey⚠

NSBlack⚠

NSBorderlessWindowMask

NSCalibratedBlackColorSpace⚠

NSCalibratedRGBColorSpace⚠

NSCalibratedWhiteColorSpace⚠

NSClipPagination

NSClosableWindowMask

NSColorPboardType⚠

NSCommandKeyMask

NSCompositeClear

NSCompositeColor

NSCompositeColorBurn

NSCompositeColorDodge

NSCompositeCopy

NSCompositeDarken

NSCompositeDestinationAtop

NSCompositeDestinationIn

NSCompositeDestinationOut

NSCompositeDestinationOver

NSCompositeDifference

NSCompositeExclusion

NSCompositeHardLight

NSCompositeHighlight

NSCompositeHue

NSCompositeLighten

NSCompositeLuminosity

NSCompositeMultiply

NSCompositeOverlay

NSCompositePlusDarker

NSCompositePlusLighter

NSCompositeSaturation

NSCompositeScreen

NSCompositeSoftLight

NSCompositeSourceAtop

NSCompositeSourceIn

NSCompositeSourceOut

NSCompositeSourceOver

NSCompositeXOR

NSControlKeyMask

NSControlStateValueMixed

NSControlStateValueOff

NSControlStateValueOn

NSControlTintDidChangeNotification⚠

NSCursorPointingDevice

NSCursorUpdate

NSCursorUpdateMask

NSCustomColorSpace⚠

NSDarkGray⚠

NSDefinitionPresentationTypeDictionaryApplication⚠

NSDefinitionPresentationTypeKey⚠

NSDefinitionPresentationTypeOverlay⚠

NSDeviceBitsPerSample⚠

NSDeviceBlackColorSpace⚠

NSDeviceCMYKColorSpace⚠

NSDeviceColorSpaceName⚠

NSDeviceIndependentModifierFlagsMask

NSDeviceIsPrinter⚠

NSDeviceIsScreen⚠

NSDeviceRGBColorSpace⚠

NSDeviceResolution⚠

NSDeviceSize⚠

NSDeviceWhiteColorSpace⚠

NSDocModalWindowMask

NSDockWindowLevel

NSDragPboard⚠

NSEraserPointingDevice

NSEventDurationForever

NSEventTrackingRunLoopMode⚠

NSFileContentsPboardType⚠

NSFilenamesPboardType⚠

NSFilesPromisePboardType⚠

NSFindPboard⚠

NSFitPagination

NSFlagsChanged

NSFlagsChangedMask

NSFloatingWindowLevel

NSFontPboard⚠

NSFontPboardType⚠

NSFullScreenModeAllScreens⚠

NSFullScreenModeApplicationPresentationOptions⚠

NSFullScreenModeSetting⚠

NSFullScreenModeWindowLevel⚠

NSFullScreenWindowMask

NSFullSizeContentViewWindowMask

NSFunctionKeyMask

NSGeneralPboard⚠

NSHTMLPboardType⚠

NSHUDWindowMask

NSHelpKeyMask

NSImageNameActionTemplate⚠

NSImageNameAddTemplate⚠

NSImageNameAdvanced⚠

NSImageNameApplicationIcon⚠

NSImageNameBluetoothTemplate⚠

NSImageNameBonjour⚠

NSImageNameBookmarksTemplate⚠

NSImageNameCaution⚠

NSImageNameColorPanel⚠

NSImageNameColumnViewTemplate⚠

NSImageNameComputer⚠

NSImageNameDotMac⚠

NSImageNameEnterFullScreenTemplate⚠

NSImageNameEveryone⚠

NSImageNameExitFullScreenTemplate⚠

NSImageNameFlowViewTemplate⚠

NSImageNameFolder⚠

NSImageNameFolderBurnable⚠

NSImageNameFolderSmart⚠

NSImageNameFollowLinkFreestandingTemplate⚠

NSImageNameFontPanel⚠

NSImageNameGoBackTemplate⚠

NSImageNameGoForwardTemplate⚠

NSImageNameGoLeftTemplate⚠

NSImageNameGoRightTemplate⚠

NSImageNameHomeTemplate⚠

NSImageNameIChatTheaterTemplate⚠

NSImageNameIconViewTemplate⚠

NSImageNameInfo⚠

NSImageNameInvalidDataFreestandingTemplate⚠

NSImageNameLeftFacingTriangleTemplate⚠

NSImageNameListViewTemplate⚠

NSImageNameLockLockedTemplate⚠

NSImageNameLockUnlockedTemplate⚠

NSImageNameMenuMixedStateTemplate⚠

NSImageNameMenuOnStateTemplate⚠

NSImageNameMobileMe⚠

NSImageNameMultipleDocuments⚠

NSImageNameNetwork⚠

NSImageNamePathTemplate⚠

NSImageNamePreferencesGeneral⚠

NSImageNameQuickLookTemplate⚠

NSImageNameRefreshFreestandingTemplate⚠

NSImageNameRefreshTemplate⚠

NSImageNameRemoveTemplate⚠

NSImageNameRevealFreestandingTemplate⚠

NSImageNameRightFacingTriangleTemplate⚠

NSImageNameShareTemplate⚠

NSImageNameSlideshowTemplate⚠

NSImageNameSmartBadgeTemplate⚠

NSImageNameStatusAvailable⚠

NSImageNameStatusNone⚠

NSImageNameStatusPartiallyAvailable⚠

NSImageNameStatusUnavailable⚠

NSImageNameStopProgressFreestandingTemplate⚠

NSImageNameStopProgressTemplate⚠

NSImageNameTouchBarAddDetailTemplate⚠

NSImageNameTouchBarAddTemplate⚠

NSImageNameTouchBarAlarmTemplate⚠

NSImageNameTouchBarAudioInputMuteTemplate⚠

NSImageNameTouchBarAudioInputTemplate⚠

NSImageNameTouchBarAudioOutputMuteTemplate⚠

NSImageNameTouchBarAudioOutputVolumeHighTemplate⚠

NSImageNameTouchBarAudioOutputVolumeLowTemplate⚠

NSImageNameTouchBarAudioOutputVolumeMediumTemplate⚠

NSImageNameTouchBarAudioOutputVolumeOffTemplate⚠

NSImageNameTouchBarBookmarksTemplate⚠

NSImageNameTouchBarColorPickerFill⚠

NSImageNameTouchBarColorPickerFont⚠

NSImageNameTouchBarColorPickerStroke⚠

NSImageNameTouchBarCommunicationAudioTemplate⚠

NSImageNameTouchBarCommunicationVideoTemplate⚠

NSImageNameTouchBarComposeTemplate⚠

NSImageNameTouchBarDeleteTemplate⚠

NSImageNameTouchBarDownloadTemplate⚠

NSImageNameTouchBarEnterFullScreenTemplate⚠

NSImageNameTouchBarExitFullScreenTemplate⚠

NSImageNameTouchBarFastForwardTemplate⚠

NSImageNameTouchBarFolderCopyToTemplate⚠

NSImageNameTouchBarFolderMoveToTemplate⚠

NSImageNameTouchBarFolderTemplate⚠

NSImageNameTouchBarGetInfoTemplate⚠

NSImageNameTouchBarGoBackTemplate⚠

NSImageNameTouchBarGoDownTemplate⚠

NSImageNameTouchBarGoForwardTemplate⚠

NSImageNameTouchBarGoUpTemplate⚠

NSImageNameTouchBarHistoryTemplate⚠

NSImageNameTouchBarIconViewTemplate⚠

NSImageNameTouchBarListViewTemplate⚠

NSImageNameTouchBarMailTemplate⚠

NSImageNameTouchBarNewFolderTemplate⚠

NSImageNameTouchBarNewMessageTemplate⚠

NSImageNameTouchBarOpenInBrowserTemplate⚠

NSImageNameTouchBarPauseTemplate⚠

NSImageNameTouchBarPlayPauseTemplate⚠

NSImageNameTouchBarPlayTemplate⚠

NSImageNameTouchBarPlayheadTemplate⚠

NSImageNameTouchBarQuickLookTemplate⚠

NSImageNameTouchBarRecordStartTemplate⚠

NSImageNameTouchBarRecordStopTemplate⚠

NSImageNameTouchBarRefreshTemplate⚠

NSImageNameTouchBarRemoveTemplate⚠

NSImageNameTouchBarRewindTemplate⚠

NSImageNameTouchBarRotateLeftTemplate⚠

NSImageNameTouchBarRotateRightTemplate⚠

NSImageNameTouchBarSearchTemplate⚠

NSImageNameTouchBarShareTemplate⚠

NSImageNameTouchBarSidebarTemplate⚠

NSImageNameTouchBarSkipAhead15SecondsTemplate⚠

NSImageNameTouchBarSkipAhead30SecondsTemplate⚠

NSImageNameTouchBarSkipAheadTemplate⚠

NSImageNameTouchBarSkipBack15SecondsTemplate⚠

NSImageNameTouchBarSkipBack30SecondsTemplate⚠

NSImageNameTouchBarSkipBackTemplate⚠

NSImageNameTouchBarSkipToEndTemplate⚠

NSImageNameTouchBarSkipToStartTemplate⚠

NSImageNameTouchBarSlideshowTemplate⚠

NSImageNameTouchBarTagIconTemplate⚠

NSImageNameTouchBarTextBoldTemplate⚠

NSImageNameTouchBarTextBoxTemplate⚠

NSImageNameTouchBarTextCenterAlignTemplate⚠

NSImageNameTouchBarTextItalicTemplate⚠

NSImageNameTouchBarTextJustifiedAlignTemplate⚠

NSImageNameTouchBarTextLeftAlignTemplate⚠

NSImageNameTouchBarTextListTemplate⚠

NSImageNameTouchBarTextRightAlignTemplate⚠

NSImageNameTouchBarTextStrikethroughTemplate⚠

NSImageNameTouchBarTextUnderlineTemplate⚠

NSImageNameTouchBarUserAddTemplate⚠

NSImageNameTouchBarUserGroupTemplate⚠

NSImageNameTouchBarUserTemplate⚠

NSImageNameTouchBarVolumeDownTemplate⚠

NSImageNameTouchBarVolumeUpTemplate⚠

NSImageNameTrashEmpty⚠

NSImageNameTrashFull⚠

NSImageNameUser⚠

NSImageNameUserAccounts⚠

NSImageNameUserGroup⚠

NSImageNameUserGuest⚠

NSInkTextPboardType⚠

NSKeyDown

NSKeyDownMask

NSKeyUp

NSKeyUpMask

NSLeftMouseDown

NSLeftMouseDownMask

NSLeftMouseDragged

NSLeftMouseDraggedMask

NSLeftMouseUp

NSLeftMouseUpMask

NSLightGray⚠

NSMainMenuWindowLevel

NSMenuDidAddItemNotification⚠

NSMenuDidBeginTrackingNotification⚠

NSMenuDidChangeItemNotification⚠

NSMenuDidEndTrackingNotification⚠

NSMenuDidRemoveItemNotification⚠

NSMenuDidSendActionNotification⚠

NSMenuItemImportFromDeviceIdentifier⚠

NSMenuWillSendActionNotification⚠

NSMiniControlSize

NSMiniaturizableWindowMask

NSMixedState

NSModalPanelRunLoopMode⚠

NSModalPanelWindowLevel

NSModalResponseAbort

NSModalResponseCancel

NSModalResponseContinue

NSModalResponseOK

NSModalResponseStop

NSMouseEntered

NSMouseEnteredMask

NSMouseEventSubtype

NSMouseExited

NSMouseExitedMask

NSMouseMoved

NSMouseMovedMask

NSMultipleTextSelectionPboardType⚠

NSNamedColorSpace⚠

NSNonactivatingPanelMask

NSNormalWindowLevel

NSNumericPadKeyMask

NSOffState

NSOnState

NSOtherMouseDown

NSOtherMouseDownMask

NSOtherMouseDragged

NSOtherMouseDraggedMask

NSOtherMouseUp

NSOtherMouseUpMask

NSPDFPboardType⚠

NSPICTPboardType⚠

NSPasteboardNameDrag⚠

NSPasteboardNameFind⚠

NSPasteboardNameFont⚠

NSPasteboardNameGeneral⚠

NSPasteboardNameRuler⚠

NSPasteboardTypeColor⚠

NSPasteboardTypeFileURL⚠

NSPasteboardTypeFindPanelSearchOptions⚠

NSPasteboardTypeFont⚠

NSPasteboardTypeHTML⚠

NSPasteboardTypeMultipleTextSelection⚠

NSPasteboardTypePDF⚠

NSPasteboardTypePNG⚠

NSPasteboardTypeRTF⚠

NSPasteboardTypeRTFD⚠

NSPasteboardTypeRuler⚠

NSPasteboardTypeSound⚠

NSPasteboardTypeString⚠

NSPasteboardTypeTIFF⚠

NSPasteboardTypeTabularText⚠

NSPasteboardTypeTextFinderOptions⚠

NSPasteboardTypeURL⚠

NSPasteboardURLReadingContentsConformToTypesKey⚠

NSPasteboardURLReadingFileURLsOnlyKey⚠

NSPatternColorSpace⚠

NSPenLowerSideMask

NSPenPointingDevice

NSPenTipMask

NSPenUpperSideMask

NSPeriodic

NSPeriodicMask

NSPopUpMenuWindowLevel

NSPostScriptPboardType⚠

NSPowerOffEventType

NSPrintAllPages⚠

NSPrintBottomMargin⚠

NSPrintCancelJob⚠

NSPrintCopies⚠

NSPrintDetailedErrorReporting⚠

NSPrintFaxNumber⚠

NSPrintFirstPage⚠

NSPrintFormName⚠

NSPrintHeaderAndFooter⚠

NSPrintHorizontalPagination⚠

NSPrintHorizontallyCentered⚠

NSPrintJobDisposition⚠

NSPrintJobFeatures⚠

NSPrintJobSavingFileNameExtensionHidden⚠

NSPrintJobSavingURL⚠

NSPrintLastPage⚠

NSPrintLeftMargin⚠

NSPrintManualFeed⚠

NSPrintMustCollate⚠

NSPrintOperationExistsException⚠

NSPrintOrientation⚠

NSPrintPagesAcross⚠

NSPrintPagesDown⚠

NSPrintPagesPerSheet⚠

NSPrintPaperFeed⚠

NSPrintPaperName⚠

NSPrintPaperSize⚠

NSPrintPreviewJob⚠

NSPrintPrinter⚠

NSPrintPrinterName⚠

NSPrintReversePageOrder⚠

NSPrintRightMargin⚠

NSPrintSaveJob⚠

NSPrintSavePath⚠

NSPrintScalingFactor⚠

NSPrintSelectionOnly⚠

NSPrintSpoolJob⚠

NSPrintTime⚠

NSPrintTopMargin⚠

NSPrintVerticalPagination⚠

NSPrintVerticallyCentered⚠

NSRTFDPboardType⚠

NSRTFPboardType⚠

NSRegularControlSize

NSResizableWindowMask

NSRightMouseDown

NSRightMouseDownMask

NSRightMouseDragged

NSRightMouseDraggedMask

NSRightMouseUp

NSRightMouseUpMask

NSRulerPboard⚠

NSRulerPboardType⚠

NSScreenChangedEventType

NSScreenSaverWindowLevel

NSScrollWheel

NSScrollWheelMask

NSShiftKeyMask

NSSmallControlSize

NSStatusWindowLevel

NSStringPboardType⚠

NSSubmenuWindowLevel

NSSystemColorsDidChangeNotification⚠

NSSystemDefined

NSSystemDefinedMask

NSTIFFPboardType⚠

NSTabletPoint

NSTabletPointEventSubtype

NSTabletPointMask

NSTabletProximity

NSTabletProximityEventSubtype

NSTabletProximityMask

NSTabularTextPboardType⚠

NSTextFinderCaseInsensitiveKey⚠

NSTextFinderMatchingTypeKey⚠

NSTexturedBackgroundWindowMask

NSTitledWindowMask

NSTornOffMenuWindowLevel

NSTouchEventSubtype

NSURLPboardType⚠

NSUnifiedTitleAndToolbarWindowMask

NSUnknownPointingDevice

NSUnscaledWindowMask

NSUtilityWindowMask

NSVCardPboardType⚠

NSViewAnimationEffectKey⚠

NSViewAnimationEndFrameKey⚠

NSViewAnimationFadeInEffect⚠

NSViewAnimationFadeOutEffect⚠

NSViewAnimationStartFrameKey⚠

NSViewAnimationTargetKey⚠

NSViewBoundsDidChangeNotification⚠

NSViewDidUpdateTrackingAreasNotification⚠

NSViewFocusDidChangeNotification⚠

NSViewFrameDidChangeNotification⚠

NSViewGlobalFrameDidChangeNotification⚠

NSWhite⚠

NSWindowDidBecomeKeyNotification⚠

NSWindowDidBecomeMainNotification⚠

NSWindowDidChangeBackingPropertiesNotification⚠

NSWindowDidChangeOcclusionStateNotification⚠

NSWindowDidChangeScreenNotification⚠

NSWindowDidChangeScreenProfileNotification⚠

NSWindowDidDeminiaturizeNotification⚠

NSWindowDidEndLiveResizeNotification⚠

NSWindowDidEndSheetNotification⚠

NSWindowDidEnterFullScreenNotification⚠

NSWindowDidEnterVersionBrowserNotification⚠

NSWindowDidExitFullScreenNotification⚠

NSWindowDidExitVersionBrowserNotification⚠

NSWindowDidExposeNotification⚠

NSWindowDidMiniaturizeNotification⚠

NSWindowDidMoveNotification⚠

NSWindowDidResignKeyNotification⚠

NSWindowDidResignMainNotification⚠

NSWindowDidResizeNotification⚠

NSWindowDidUpdateNotification⚠

NSWindowExposedEventType

NSWindowFullScreenButton

NSWindowMovedEventType

NSWindowWillBeginSheetNotification⚠

NSWindowWillCloseNotification⚠

NSWindowWillEnterFullScreenNotification⚠

NSWindowWillEnterVersionBrowserNotification⚠

NSWindowWillExitFullScreenNotification⚠

NSWindowWillExitVersionBrowserNotification⚠

NSWindowWillMiniaturizeNotification⚠

NSWindowWillMoveNotification⚠

NSWindowWillStartLiveResizeNotification⚠

Traits

CIColorNSAppKitAdditions

Category “NSAppKitAdditions” on [CIColor].

NSAccessibility

NSAccessibilityButton

NSAccessibilityCheckBox

NSAccessibilityContainsTransientUI

NSAccessibilityElementLoading

NSAccessibilityElementProtocol

NSAccessibilityGroup

NSAccessibilityImage

NSAccessibilityLayoutArea

NSAccessibilityLayoutItem

NSAccessibilityList

NSAccessibilityNavigableStaticText

NSAccessibilityOutline

NSAccessibilityProgressIndicator

NSAccessibilityRadioButton

NSAccessibilityRow

NSAccessibilitySlider

NSAccessibilityStaticText

NSAccessibilityStepper

NSAccessibilitySwitch

NSAccessibilityTable

NSAnimatablePropertyContainer

NSAnimationDelegate

NSAppearanceCustomization

NSApplicationDelegate

NSBundleImageExtension

Category on NSBundle.

NSDraggingDestination

NSDraggingInfo

NSDraggingSource

NSImageDelegate

NSMenuDelegate

NSMenuItemValidation

NSOpenSavePanelDelegate

NSPasteboardReading

NSPasteboardTypeOwner

NSPasteboardWriting

NSServicesMenuRequestor

NSSpringLoadingDestination

NSStandardKeyBindingResponding

NSTextFinderBarContainer

NSTextFinderClient

NSURLNSPasteboardSupport

Category “NSPasteboardSupport” on NSURL.

NSUserInterfaceItemIdentification

NSUserInterfaceValidations

NSValidatedUserInterfaceItem

NSViewLayerContentScaleDelegate

NSViewToolTipOwner

NSWindowDelegate

Functions

NSApp

NSApplicationLoad⚠

NSApplicationMain⚠

NSAvailableWindowDepths⚠

NSBeep⚠

NSBestDepth⚠

NSBitsPerPixelFromDepth⚠

NSBitsPerSampleFromDepth⚠

NSColorSpaceFromDepth⚠

NSCopyBits⚠Deprecated

NSCountWindows⚠Deprecated

NSCountWindowsForContext⚠Deprecated

NSCreateFileContentsPboardType⚠

NSCreateFilenamePboardType⚠

NSDisableScreenUpdates⚠Deprecated

NSDottedFrameRect⚠

NSDrawButton⚠

NSDrawColorTiledRects⚠

NSDrawDarkBezel⚠

NSDrawGrayBezel⚠

NSDrawGroove⚠

NSDrawLightBezel⚠

NSDrawNinePartImage⚠

NSDrawThreePartImage⚠

NSDrawTiledRects⚠

NSDrawWhiteBezel⚠

NSDrawWindowBackground⚠

NSEnableScreenUpdates⚠Deprecated

NSEraseRect⚠

NSFrameRect⚠

NSFrameRectWithWidth⚠

NSFrameRectWithWidthUsingOperation⚠

NSGetFileType⚠

NSGetFileTypes⚠

NSGetWindowServerMemory⚠Deprecated

NSHighlightRect⚠Deprecated

NSNumberOfColorComponents⚠

NSPerformService⚠

NSPlanarFromDepth⚠

NSReadPixel⚠Deprecated

NSRectClip⚠

NSRectClipList⚠

NSRectFill⚠

NSRectFillList⚠

NSRectFillListUsingOperation⚠

NSRectFillListWithColors⚠

NSRectFillListWithColorsUsingOperation⚠

NSRectFillListWithGrays⚠

NSRectFillUsingOperation⚠

NSRegisterServicesProvider⚠

NSReleaseAlertPanel⚠Deprecated

NSSetFocusRingStyle⚠

NSSetShowsServicesMenuItem⚠

NSShowAnimationEffect⚠Deprecated

NSShowsServicesMenuItem⚠

NSUnregisterServicesProvider⚠

NSUpdateDynamicServices⚠

NSWindowList⚠Deprecated

NSWindowListForContext⚠Deprecated

Type Aliases

NSAboutPanelOptionKey

NSAnimatablePropertyKey

NSAnimationProgress

NSAppKitVersion

NSAppearanceName

NSCellStateValue

NSColorSpaceName

NSControlStateValue

NSDefinitionOptionKey

NSDefinitionPresentationType

NSDeviceDescriptionKey

NSImageName

NSModalResponse

NSModalSession

NSPasteboardName

NSPasteboardReadingOptionKey

NSPasteboardType

NSPasteboardTypeTextFinderOptionKey

NSPrintInfoAttributeKey

NSPrintInfoSettingKey

NSPrintJobDispositionValue

NSServiceProviderName

NSToolTipTag

NSTrackingRectTag

NSUserInterfaceItemIdentifier

NSViewAnimationEffectName

NSViewAnimationKey

NSViewFullScreenModeOptionKey

NSWindowFrameAutosaveName

NSWindowLevel

NSWindowPersistableFrameDescriptor

NSWindowTabbingIdentifier